Restructured quantum folder
