--- pymuonsuite/calculate/nuclearquantum/__init__.py
@@ -1,3 +1,0 @@
-"""
-Author: Adam Laverack
-"""
