+++ pymuonsuite/quantum/__init__.py
@@ -1,0 +1,5 @@
+"""
+Methods to approximate quantum nuclear effects for muon
+
+Author: Adam Laverack, Simone Sturniolo
+"""
