+++ pymuonsuite/quantum/__main__.py
@@ -10,22 +10,23 @@ from __future__ import unicode_literals
 
 import argparse as ap
 
-from pymuonsuite.calculate.nuclearquantum.phonon import phonon_hfcc
+from pymuonsuite.calculate.quantum.phonons.castep import phonon_hfcc
 from pymuonsuite.schemas import load_input_file, PhononHfccSchema
+
 
 def nq_entry():
     parser = ap.ArgumentParser()
     parser.add_argument('calculation_type', type=str,
-        help="""Type of calculation to be performed, currently supports:
+                        help="""Type of calculation to be performed, currently supports:
                 'phonon_hfcc': Nuclear quantum effects simulated by phonons""")
     parser.add_argument('parameter_file', type=str,
-        help="YAML file containing relevant input parameters")
+                        help="YAML file containing relevant input parameters")
     parser.add_argument('-w',   action='store_true', default=False,
                         help="Create and write input files instead of parsing the results")
 
     args = parser.parse_args()
 
-    #Load parameters
+    # Load parameters
     params = load_input_file(args.parameter_file, PhononHfccSchema)
 
     if args.calculation_type == "phonon_hfcc":
@@ -33,6 +34,7 @@ def nq_entry():
     else:
         raise RuntimeError("""Invalid calculation type entered, please use
                               python -h flag to see currently supported types""")
+
 
 if __name__ == "__main__":
     nq_entry()
