+++ pymuonsuite/quantum/phonons/__init__.py
@@ -1,0 +1,4 @@
+"""
+Methods to extract or compute phonon modes for muons and other atoms for use
+in nuclear quantum effect approximations.
+"""
